Add defaults to config and update install README
Former-commit-id: 4a05eabce300fc7a3862e644c640b27ed59b9f7b [formerly e83316fe79a1042f2eb1d80f8efd15dfee2bcaad]
Former-commit-id: 5638e06e9be69e81b926bfd0bbf36da0ac94d4b9
